requirements.txt
keras==2.10.0
numpy==1.24.1
pandas==1.5.2
plotly==5.13.0
PyYAML==6.0
scikit-learn==1.2.1
streamlit==1.17.0
streamlit-authenticator==0.2.1
tensorflow==2.10.1